Fix: sync .scrapy folder only if it exists

+++ sync.sh
@@ -3,7 +3,9 @@ set -xe
 MESSAGE=$1
 
 # snapshot scrapy cache
-aws s3 cp .scrapy s3://player-scores/scrapy-httpcache/ --recursive
+if [ -e .scrapy ]; then
+  aws s3 cp .scrapy s3://player-scores/scrapy-httpcache/ --recursive
+fi
 
 # snapshot prepared files
 aws s3 cp data/prep s3://player-scores/snapshots/prep/ --recursive --exclude "*.json"
